Move security related build steps into security package with own processor
