+++ commons/deployment/src/main/java/com/github/mcollovati/quarkus/hilla/deployment/security/AuthFormBuildItem.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.mcollovati.quarkus.hilla.deployment;
+package com.github.mcollovati.quarkus.hilla.deployment.security;
 
 import io.quarkus.builder.item.SimpleBuildItem;
