+++ commons/runtime/src/main/java/com/github/mcollovati/quarkus/hilla/security/QuarkusNavigationAccessControl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.mcollovati.quarkus.hilla;
+package com.github.mcollovati.quarkus.hilla.security;
 
 import jakarta.enterprise.event.Observes;
 import jakarta.inject.Inject;
